pyproject.toml
[tool.poetry]
name = "bookshelf"
version = "0.1.0"
description = "An application to manage your library"
authors = ["Christophe Bastin <5850561+bastinc@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.6.2"
Django = "2.2.24"
python-dotenv = "^0.19.1"
black = "^21.9b0"
flake8 = "^4.0.1"
django-oauth-toolkit = "1.4.1"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.15.0"
commitizen = "^2.20.0"
pytest-dotenv = "^0.5.2"
pytest-django = "^4.4.0"
factory-boy = "^3.2.0"


[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.1"
tag_format = "$version"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
